% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mln_generator.R
\name{mln_generator}
\alias{mln_generator}
\title{Generate multilayer networks from multilayer stochastic block models
and multilayer stochastic co-block models.}
\usage{
mln_generator(row.label, col.label, probmat, directed = FALSE, sparse = TRUE)
}
\arguments{
\item{row.label}{The row community vector (from \code{1:kr}) with the numbers indicating which row community each node is assigned.}

\item{col.label}{The column community vector (from \code{1:kc}) with the numbers indicating which column community each node is assigned.
\code{row.label} and \code{col.label} must have the same length but not necessarily indicate the same communities and community numbers.
Only required when \code{directed = TRUE}.}

\item{probmat}{List of link probability matrices with dimension \code{c(kr, kc)}.
\code{probmat[l][i,j]} is proportional to the probability of an edge from nodes in row community \code{i} to nodes in column community \code{j} in layer \code{l}.}

\item{directed}{If \code{TRUE}, the multilayer directed networks is generated from the multilayer stochastic co-block model.
If \code{FALSE}, the multilayer undirected networks is generated from the multilayer stochastic block model. Default is \code{FALSE}.}

\item{sparse}{Logical indicating whether to return sparse matrices (\code{dgCMatrix}) for network adjacency matrices.
Default is \code{TRUE}.}
}
\value{
A list of adjacency matrices representing the generated multilayer networks.
}
\description{
Generate adjacency matrices of multilayer undirected networks from
multilayer stochastic block models and multilayer directed networks from
multilayer stochastic co-block models.
}
\details{
This function generates multilayer undirected (directed) networks using
multilayer stochastic block (co-block) models.In the multilayer stochastic
block (co-block) models, all layers share common row and column communities
but with possibly different edge densities. The layer-wise networks are
generated independently from the stochastic block (co-block) model.
}
\examples{
# The multilayer stochastic block model
n <- 300
kr <- 2
row.label <- sample(rep(1:kr, each = n/kr), n)
probmat1 <- matrix(0.05, kr, kr) + diag(c(0.1, 0.05))
probmat2 <- matrix(0.01, kr, kr) + diag(c(0.05, 0.02))
probmats <- list(probmat1, probmat2, probmat2)
mln1 <- mln_generator(row.label = row.label, probmat = probmats, directed = FALSE)

# The multilayer stochastic co-block model
n <- 300
kr <- 2
kc <- 3
row.label <- sample(rep(1:kr, each = n/kr), n)
col.label <- sample(rep(1:kc, each = n/kc), n)
probmat1 <- matrix(0.05, kr, kc)
probmat1[1,1] <- 0.1
probmat1[2,2] <- 0.1
probmat2 <- matrix(0.01, kr, kc)
probmat1[1,1] <- 0.05
probmat1[2,3] <- 0.03
probmats <- list(probmat1, probmat2, probmat2)
mln2 <- mln_generator(row.label, col.label, probmats, directed = TRUE)

}
\references{
W. Su, X. Guo, X. Chang and Y. Yang. (2024)
\emph{Spectral co-clustering in multi-layer directed networks},
\emph{Computational Statistics & Data Analysis, Vol. 198, 107987}\cr
\url{https://doi.org/10.1016/j.csda.2024.107987}\cr
}
